// src/app/api-client/model/groupFamilyListResponse.ts
/**
 * FastAPI
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GroupFamilyResponse } from './groupFamilyResponse';

/**
 * Response listing all related-group families.
 */
export interface GroupFamilyListResponse {
  families: Array<GroupFamilyResponse>;
}