// src/app/api-client/model/groupAliasListResponse.ts
/**
 * Horde Model Reference API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GroupAliasResponse } from './groupAliasResponse';

/**
 * Response listing all group alias entries.
 */
export interface GroupAliasListResponse {
  entries: Array<GroupAliasResponse>;
}