// src/app/api-client/model/detectFamiliesResponse.ts
/**
 * FastAPI
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GroupFamilyResponse } from './groupFamilyResponse';

/**
 * Response from auto-detection of family suggestions.
 */
export interface DetectFamiliesResponse {
  suggestions: Array<GroupFamilyResponse>;
  total_groups_analyzed: number;
  groups_in_families: number;
  standalone_groups: number;
}